+++ CHANGELOG.md
@@ -12,6 +12,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.1] - 2025-12-10
 
-### Added
+### Changed
 - Decline transaction note updated
+
+## [1.0.2] - 2025-01-09
+
+### Changed
+- Expiry date input validation updated
+
+### Added
+- Customer legal entity
+